Split eval and pattern_eval
