+++ src/book.rs
@@ -192,7 +192,13 @@ fn search<Eval: Evaluator>(
             node_count: 0,
             cache_gen: solve_obj.cache_gen,
         };
-        let (_score, hand, _depth) = think_parallel(&searcher, board, EVAL_SCORE_MIN, EVAL_SCORE_MAX, false);
+        let (_score, hand, _depth) = think_parallel(
+            &searcher,
+            board,
+            solve_obj.evaluator.score_min(),
+            solve_obj.evaluator.score_max(),
+            false,
+        );
         hand
     }
 }
