+++ src/main.rs
@@ -1,5 +1,6 @@
 #![feature(const_option)]
 #![feature(portable_simd)]
+#![feature(iterator_try_collect)]
 #![feature(test)]
 mod book;
 mod compression;
