+++ src/setup.rs
@@ -1,4 +1,4 @@
-use crate::engine::eval::*;
+use crate::engine::pattern_eval::*;
 use crate::engine::search::*;
 use crate::engine::table::*;
 use std::path::Path;
