+++ src/train.rs
@@ -2,6 +2,7 @@ use crate::engine::bits::*;
 use crate::engine::board::*;
 use crate::engine::eval::*;
 use crate::engine::hand::*;
+use crate::engine::pattern_eval::*;
 use crate::engine::table::*;
 use crate::engine::think::*;
 use crate::sparse_mat::*;
@@ -464,8 +465,8 @@ pub fn eval_stats(matches: &ArgMatches) -> Option<()> {
             for depth in 1..=depth_max {
                 if let Some((evaluated, _)) = searcher.think(
                     board,
-                    EVAL_SCORE_MIN,
-                    EVAL_SCORE_MAX,
+                    evaluator.score_min(),
+                    evaluator.score_max(),
                     false,
                     depth as i32 * DEPTH_SCALE,
                 ) {
@@ -502,7 +503,11 @@ pub fn eval_stats(matches: &ArgMatches) -> Option<()> {
             for idx in 1..=15 {
                 let ratio = 1.0 - 0.7f32.powi(idx);
                 let index = (total as f32 * ratio) as usize;
-                println!("{} {}", ratio, vd[index] as f32 / SCALE as f32);
+                println!(
+                    "{} {}",
+                    ratio,
+                    vd[index] as f32 / evaluator.score_scale() as f32
+                );
             }
         }
     }
